% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{AgevsProter_Canada_full}
\alias{AgevsProter_Canada_full}
\title{Population du Canada par classes d’âge et provinces et territoires en 2020}
\format{
Un jeu de données avec 21 observations de 13 variables.
\describe{
\item{Terre.Neuve.et.Labrador}{}
\item{Île.du.Prince.Édouard}{}
\item{Nouvelle.Écosse}{}
\item{Nouveau.Brunswick}{}
\item{Québec}{}
\item{Ontario}{}
\item{Manitoba}{}
\item{Saskatchewan}{}
\item{Alberta}{}
\item{Colombie.Britannique}{}
\item{Yukon}{}
\item{TerritoiresduNord.Ouest}{}
\item{Nunavut}{}
}
}
\usage{
AgevsProter_Canada_full
}
\description{
Répartition en classes d’âge de la population des provinces et des territoires du Canada en 2020. Statistique Canada. Tableau 17-10-0005-01  Estimations de la population au 1er juillet, par âge et sexe.
}
\references{
\url{https://doi.org/10.25318/1710000501-fra}
}
\keyword{datasets}
