% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{att_rep_ind}
\alias{att_rep_ind}
\title{Indices d'attraction/répulsion}
\usage{
att_rep_ind(data)
}
\arguments{
\item{data}{Jeux de données}
}
\value{
Liste à un élément qui content le tableau croisé des indices.
}
\description{
Fonction de calcul et de représentation des indices d'attraction/répulsion
}
\examples{

data(champignons)
champ_sel <- champignons[,c("couleur_chapeau","contusions","odeur",
"espacement_lamelle","habitat")]
sageR::att_rep_ind(champ_sel)

}
