% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bitcoin}
\alias{bitcoin}
\title{Prix journalier du Bitcoin du 31/12/2014 au 15/05/2018}
\format{
A data frame with 1233 observations on the following 2 variables.
\describe{
\item{Date}{a POSIXct}
\item{Bitcoin}{a numeric vector}
}
}
\source{
Yahoo finance (https://fr.finance.yahoo.com/quote/BTC-EUR/history?p=BTC-EUR)
}
\usage{
bitcoin
}
\description{
Les données décrivent l'évolution du prix journalier du Bitcoin sur la période du
31/12/2014 au 15/05/2018. Les données sont publiques et disponibles sur le site de
Yahoo finance (https://fr.finance.yahoo.com/quote/BTC-EUR/history?p=BTC-EUR).
}
\examples{

data(bitcoin)
str(bitcoin)
plot(bitcoin)

}
\keyword{datasets}
