% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{champignons}
\alias{champignons}
\title{Caractéristiques de champignons}
\format{
A data frame with 1000 observations on the following 8 variables.
\describe{\item{num_champ}{a numeric vector. Numéro du champignon.}
\item{classes}{a factor with levels \code{comestible} \code{veneneux}. Classes.}
\item{couleur_chapeau}{a factor with levels \code{blanc} \code{gris} \code{jaune} \code{marron} \code{rouge}. Couleur du chapeau.}
\item{contusions}{a factor with levels \code{abime} \code{non}. Contusions.}
\item{odeur}{a factor with levels \code{amande} \code{anis} \code{apre} \code{inodore}. Odeur du champignon.}
\item{espacement_lamelle}{a factor with levels \code{ferme} \code{serre}. Espacement des lamelles.}
\item{couleur_tige_dessus}{a factor with levels \code{blanc} \code{gris} \code{rose}. Couleur du dessus de la tige.}
\item{habitat}{a factor with levels \code{foret} \code{pelouse} \code{prairies} \code{sentier} \code{urbain}. Habitat.}
}
}
\usage{
champignons
}
\description{
Caractéristiques de 1000 champignons.
}
\examples{

data(champignons)
str(champignons)
summary(champignons)

}
\keyword{datasets}
