% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Copies}
\alias{Copies}
\title{Correction de copies}
\format{
Un jeu de données avec 30 observations de 2 variables.
\describe{
\item{Correcteur A}{}
\item{Correcteur B}{}
}
}
\usage{
Copies
}
\description{
Notes de deux correcteurs (A et B) pour les mêmes trente copies.
}
\keyword{datasets}
