% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Essence}
\alias{Essence}
\title{Prix de l'essence sans plomb 95}
\format{
Un jeu de données avec 60 observations de 2 variables.
\describe{
\item{Aube}{}
\item{Marne}{}
}
}
\usage{
Essence
}
\description{
Prix de l'essence sans plomb 95 en avril 2021 dans deux départements français, l'aube et la marne.
}
\keyword{datasets}
