% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{funcCircleCoords}
\alias{funcCircleCoords}
\title{Calcule les coordonnées des points d'un cercle (Generate circle coordinates)}
\source{
Adapted from Joran's response to \url{https://stackoverflow.com/questions/6862742/draw-a-circle-with-ggplot2}.
}
\usage{
funcCircleCoords(center = c(0, 0), r = 1, npoints = 100)
}
\arguments{
\item{center}{coordinate for centroid}

\item{r}{radius}

\item{npoints}{number of coordinates to generate}
}
\value{
a dataframe
}
\description{
Generate coordinates to draw a circle.
}
\examples{
funcCircleCoords(c(1,2),1)
plot(funcCircleCoords(c(1,2),1))
}
