% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.R
\name{addKey,SummarizedExperiment-method}
\alias{addKey,SummarizedExperiment-method}
\alias{addKey}
\alias{filterKey}
\alias{getKey}
\alias{addKey,Spectra-method}
\alias{addKey,PSM-method}
\alias{addKey,QFeatures-method}
\alias{filterKey,Spectra-method}
\alias{filterKey,PSM-method}
\alias{filterKey,SummarizedExperiment-method}
\alias{filterKey,QFeatures-method}
\alias{filterKey,MsExperiment-method}
\alias{getKey,PSM-method}
\alias{getKey,Spectra-method}
\alias{getKey,SummarizedExperiment-method}
\alias{getKey,QFeatures-method}
\title{Add and subsetting keys}
\usage{
\S4method{addKey}{SummarizedExperiment}(object, vars = NULL, key = ".KEY", force = FALSE, sep = ".")

\S4method{addKey}{Spectra}(object, vars = NULL, key = ".KEY", force = FALSE, sep = ".")

\S4method{addKey}{PSM}(object, vars = NULL, key = ".KEY", force = FALSE, sep = ".")

\S4method{addKey}{QFeatures}(object, vars = NULL, key = ".KEY", force = FALSE, sep = ".")

\S4method{filterKey}{Spectra}(object, value, key = ".KEY")

\S4method{filterKey}{PSM}(object, value, key = ".KEY")

\S4method{filterKey}{SummarizedExperiment}(object, value, key = ".KEY")

\S4method{filterKey}{QFeatures}(object, value, key = ".KEY", keep = FALSE)

\S4method{filterKey}{MsExperiment}(
  object,
  value,
  key = ".KEY",
  data = c("spectra", "qdata"),
  otherdata = NULL,
  keep = FALSE
)

\S4method{getKey}{PSM}(object, key = ".KEY")

\S4method{getKey}{Spectra}(object, key = ".KEY")

\S4method{getKey}{SummarizedExperiment}(object, key = ".KEY")

\S4method{getKey}{QFeatures}(object, key = ".KEY", drop = TRUE)
}
\arguments{
\item{object}{An instance of class \code{SummarizedExperiment},
\code{QFeatures}, \code{PSM} or \code{Spectra}.}

\item{vars}{\code{character()} referencing the variables to use to
create the key. These are to be found in the object's
\code{rowData} (for \code{SummarizedExperiment} and \code{QFeatures} objects)
or \code{spectraData} (for \code{Spectra} objects) or in the object
itself (for \code{PSM} objects). Throws an error if the variables
aren't found. Default is \code{NULL}.}

\item{key}{\code{character(1)} defining the key's name. Default is
\code{".KEY"}.}

\item{force}{\code{logical(1)} that defines is an existing key should
be overwritten. Default is \code{FALSE} and an error is thrown if
the key is already present.}

\item{sep}{\code{character(1)} specifying the separator when multiple
variables are pasted together to generate the key. Default is
\code{"."}.}

\item{value}{\code{character()} containing the value of key of the
feature(s) of interest.}

\item{keep}{\code{logical(1)} that defines whether to keep assays that
didn't contain the key of interest. If \code{TRUE}, these are
returned as empty assays (i.e. with zero
rows/features)\verb{. Default is }FALSE`; only assays containing
matching keys are returned.}

\item{data}{\code{character()} defining what data type to subset by
key. Default are \code{"spectra"} and \code{"qdata"}. Note that these
data are ignored if their slot is NULL.}

\item{otherdata}{\code{character()} listing the name(s) of the
additional data types in \code{otherData(.)} to run \code{filterKey()}
on.}

\item{drop}{\code{logical(1)} defining whether to drop assays that
don't have the key variable. Default is \code{TRUE}.}
}
\description{
A key is a variable that is used to identify features within an
object. These features can refer to spectra in a \code{Spectra} object,
peptide-spectrum matches in a \code{PSM} object, or PSMs, peptides of
proteins in a \code{SummarizedExperiment} or \code{QFeatures} object. They
do not need to be unique.

The goal of a key is to identify matching features across data
types, such as for example to match de MS spectra to the PSMs and
peptides. These different data types can be stored together in an
\code{MsExperiment} object, and matched through one or multiple keys.

There are two general functions to work with keys across objects:
\itemize{
\item \code{addKey(object, vars, key, force, sep)} will add a key named
\code{key} to an object based on user-defined variables \code{vars}.
\item \code{filterKey(object, value, key, keep)} will subset \code{object}
based on a \code{value} in \code{key}.
\item \code{getKey(object, key, drop)} will return the values of variable
\code{key} in \code{object}.
}
}
