% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Marque.Valeur.large}
\alias{Marque.Valeur.large}
\title{Prix de vente en fonction de la marque, format large ou dépilé}
\format{
Un jeu de données avec 30 observations de 3 variables.
\describe{
\item{Marque.1}{}
\item{Marque.2}{}
\item{Marque.3}{}
}
}
\usage{
Marque.Valeur.large
}
\description{
Valeurs d'articles de qualité équivalente en fonction de leur marque.
}
\keyword{datasets}
