% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Marque.Valeur}
\alias{Marque.Valeur}
\title{Prix de vente en fonction de la marque, format long ou empilé}
\format{
Un jeu de données avec 90 observations de 2 variables.
\describe{
\item{Marque}{}
\item{Valeur}{}
}
}
\usage{
Marque.Valeur
}
\description{
Valeurs d'articles de qualité équivalente en fonction de leur marque.
}
\keyword{datasets}
