% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Precipitations_USA}
\alias{Precipitations_USA}
\title{Précipitations aux USA en 1975}
\format{
Un jeu de données avec 60 observations de 4 variables.
\describe{
\item{Ville}{}
\item{Precipitation..inches.}{}
\item{Precipitation..cms.}{}
\item{Etat}{}
}
}
\usage{
Precipitations_USA
}
\description{
Précipitations relevées dans soixante villes aux États-Unis d'Amérique en 1975.
}
\keyword{datasets}
