% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Proter_Canada}
\alias{Proter_Canada}
\title{Population du Canada par provinces et territoires en 2020}
\format{
A data frame with 13 observations on the following 2 variables.
\describe{
\item{ProTer}{a character vector}
\item{Population}{a numeric vector}
}
}
\usage{
Proter_Canada
}
\description{
Répartition de la population des provinces et des territoires du Canada en 2020. Statistique Canada. Tableau 17-10-0005-01  Estimations de la population au 1er juillet, par âge et sexe.
}
\examples{

data(Proter_Canada)
str(Proter_Canada)
barplot(Proter_Canada$Population, names.arg=Proter_Canada$ProTer, las=2)

}
\references{
\doi{10.25318/1710000501-fra}
}
\keyword{datasets}
