% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ReaFull}
\alias{ReaFull}
\title{Personnes en réanimation atteintes de la Covid 19 (21/02/2021)}
\format{
Un jeu de données avec 19 observations de 12 variables.
\describe{
\item{Région}{}
\item{Tous.âges}{}
\item{X0.9}{}
\item{X11.19}{}
\item{X20.29}{}
\item{X30.39}{}
\item{X40.49}{}
\item{X50.59}{}
\item{X60.69}{}
\item{X70.79}{}
\item{X80.89}{}
\item{X90.}{}
}
}
\usage{
ReaFull
}
\description{
Répartition par région française du nombre de personne en réanimation et atteintes de la Covid 19 le 21 février 2021.
}
\keyword{datasets}
