% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Resistance}
\alias{Resistance}
\title{Résistance à l'éclatement}
\format{
Un vecteur avec 30 observations.
}
\usage{
Resistance
}
\description{
Résistance à l'éclatement.
}
\keyword{datasets}
