% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Richesse}
\alias{Richesse}
\title{Part du revenu national total équivalent	en Euro en 2019}
\format{
Un jeu de données avec 10 observations de 36 variables.
\describe{
\item{"Déciles"}{}
\item{"Belgique"}{}
\item{"Bulgarie"}{}
\item{"Tchéquie"}{}
\item{"Danemark"}{}
\item{"Allemagne..jusqu.en.1990..ancien.territoire.de.la.RFA."}{}
\item{"Estonie"}{}
\item{"Irlande"}{}
\item{"Grèce"}{}
\item{"Espagne"}{}
\item{"France"}{}
\item{"Croatie"}{}
\item{"Italie"}{}
\item{"Chypre"}{}
\item{"Lettonie"}{}
\item{"Lituanie"}{}
\item{"Luxembourg"}{}
\item{"Hongrie"}{}
\item{"Malte"}{}
\item{"Pays.Bas"}{}
\item{"Autriche"}{}
\item{"Pologne"}{}
\item{"Portugal"}{}
\item{"Roumanie"}{}
\item{"Slovénie"}{}
\item{"Slovaquie"}{}
\item{"Finlande"}{}
\item{"Suède"}{}
\item{"Islande"}{}
\item{"Norvège"}{}
\item{"Suisse"}{}
\item{"Royaume.Uni"}{}
\item{"Monténégro"}{}
\item{"Macédoine.du.Nord"}{}
\item{"Serbie"}{}
\item{"Turquie"}{}
}
}
\usage{
Richesse
}
\description{
Répartitition du revenu par quantiles - enquêtes EU-SILC et PCM (ILC_DI01).
}
\keyword{datasets}
