% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sageMsExperiment.R
\name{sageMsExperiment}
\alias{sageMsExperiment}
\title{Import raw and sage data}
\usage{
sageMsExperiment(rawData, quantData, idData)
}
\arguments{
\item{rawData}{\code{character(1)} containing the path to the
quantification raw MS data files or a \code{\link[Spectra:Spectra]{Spectra::Spectra()}}
object containing the raw MS data.}

\item{quantData}{\code{character(1)} containing the path to the
quantification result file (typically "quant.tsv"), or an
object or class \code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}}
or \code{\link[QFeatures:QFeatures-class]{QFeatures::QFeatures()}}.}

\item{idData}{\code{character(1)} containing the path to the
identification results (typically "results.sage.tsv"), or an
instance of class \code{\link[PSMatch:PSM]{PSMatch::PSM()}}.}
}
\value{
An instance of class \code{\link[MsExperiment:MsExperiment]{MsExperiment::MsExperiment()}}.
}
\description{
This function imports sage quantitation and identification results
and the raw data that produced them and combines all into a
\code{\link[MsExperiment:MsExperiment]{MsExperiment::MsExperiment()}} object.
}
\author{
Laurent Gatto
}
