% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagePSM.R
\name{sagePSM}
\alias{sagePSM}
\title{Import sage identification results}
\usage{
sagePSM(
  idFile,
  spectrum = "scannr",
  peptide = "peptide",
  protein = "proteins",
  decoy = "label",
  rank = "rank",
  score = "hyperscore",
  ...
)
}
\arguments{
\item{idFile}{\code{character(1)} containing the identification
results, typically "results.sage.tsv".}

\item{spectrum}{\code{character(1)} variable name that defines a
spectrum in the PSM data. Default are \code{"scannr"}.}

\item{peptide}{\code{character(1)} variable name that defines a peptide
in the PSM data. Detaults are \code{"peptide"}.}

\item{protein}{protein \code{character(1)} variable name that defines a
protein in the PSM data. Detaults are \code{"proteins"}.}

\item{decoy}{\code{character(1)} variable name that defines a decoy hit
in the PSM data. Detaults are \code{"label"}.}

\item{rank}{rank \code{character(1)} variable name that defines the
rank of the peptide spectrum match in the PSM data. Default is
\code{"rank"}.}

\item{score}{score \code{character(1)} variable name that defines the
PSM score. default is \code{"hyperscore"}.}

\item{...}{Additional arguments passed to \code{\link[=read.delim]{read.delim()}}.}
}
\value{
An instance of class \code{\link[=PSM]{PSM()}}.
}
\description{
This function imports sage identification results as a
\code{\link[PSMatch:PSM]{PSMatch::PSM()}} object.
}
\examples{

idFile <- BiocFileCache::bfcquery(
                             BiocFileCache::BiocFileCache(),
                             "sageRes")$fpath
sagePSM(idFile)
}
\author{
Laurent Gatto
}
