% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sageQFeatures.R
\name{sageQFeatures}
\alias{sageQFeatures}
\title{Import sage results}
\usage{
sageQFeatures(
  quantFile,
  idFile,
  byQuant = c("file", "scannr"),
  byId = c("filename", "scannr"),
  splitBy = byQuant[1],
  quantPattern = "tmt_",
  class = c("SummarizedExperiment", "SingleCellExperiment"),
  ...
)
}
\arguments{
\item{quantFile}{\code{character(1)} containing the quantification
results, typically "quant.tsv".}

\item{idFile}{\code{character(1)} containing the identification
results, typically "results.sage.tsv".}

\item{byQuant}{\code{character()} containing the specifications of the
quantification columns used for merging. The first element
should correspond the the raw files. Passed the \code{\link[=merge]{merge()}} as
\code{by.x}. Default is \code{c("file", "scannr")}.}

\item{byId}{\code{character()} containing the specifications of the
identification columns used for merging. Passed the \code{\link[=merge]{merge()}}
as \code{by.y}. Default is \code{c("filename", "scannr")}.}

\item{splitBy}{A character refering the a \code{quantFile} or \code{idFile}
column that will be used to split the merged table into
different assays, i.e. acquisitions corresponding to different
sets of samples. Default is to use \code{byQuant[1]}. Set to \code{NULL}
to not split and return a result with a single assay, for
instance if several fractions from one sample were used.}

\item{quantPattern}{\code{character(1)} defining the pattern passed to
\code{\link[=grep]{grep()}} to extract the columns containing quantitative data.}

\item{class}{\code{character(1)} with pne of \code{"SummarizedExperiment"}
or \code{"SingleCellExperiment"} defining the assay's
class. Default is the former.}

\item{...}{Additional parameters passed to \code{\link[=read.delim]{read.delim()}}.}
}
\value{
An instance of class \code{\link[=QFeatures]{QFeatures()}} with as many assays as
defined by \code{splitBy}.
}
\description{
This function imports sage TMT quantitation and identification
results and combines them into a
\code{\link[QFeatures:QFeatures-class]{QFeatures::QFeatures()}} object.
}
\examples{

idFile <- BiocFileCache::bfcquery(
                             BiocFileCache::BiocFileCache(),
                             "sageRes")$fpath
quantFile <- BiocFileCache::bfcquery(
                                BiocFileCache::BiocFileCache(),
                                "sageQuant")$fpath

sageQFeatures(quantFile, idFile)
}
\author{
Laurent Gatto
}
