% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagerData.R
\name{sager_rids}
\alias{sager_rids}
\title{Internal functions}
\usage{
sager_rids(which = c("quant", "id", "mzml", "config"))
}
\arguments{
\item{which}{\code{character()} specifying what type of resource to
return. One or multiple of \code{"quant"}, \code{"id"} and \code{"mzml"}.

ChangeLog:
\itemize{
\item version 2: updated 'subset' data files and new config file
\item version 3: provide 3 separate subsetted mzML files, and update
quant and id files (generated from re-running sage on the mzML
subsets).
}}
}
\description{
\itemize{
\item \code{sager_rids()} is used internally to keep track of the resource
names (rname) in the package's cache. It is used in functions
such as \code{\link[=sagerQuantData]{sagerQuantData()}}, ... that return the path to a
resource (rpath). When new data are added to the cache, their
rname should be added here.
\item \code{sager_urls()} keeps track of the resource URLs on zenodo. It is
used in \code{\link[=sagerAddData]{sagerAddData()}} the get the URLs of the resources to be
added.
}

Both return named vectors.
}
