% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagerData.R
\name{sager_rids}
\alias{sager_rids}
\title{Sager Cache Resource Identifiers}
\usage{
sager_rids(which = c("quant", "id", "mzml"))
}
\arguments{
\item{which}{\code{character()} specifying what type of resource to
return.}
}
\description{
This function is used internally to keep track of the resource
names (rname) in the package's cache in functions such as
\code{sagerQuantData()}, ... that return the path to a resource
(rpath). When new data are added to the cache, their rname should
be added here.
}
