% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sagerData}
\alias{sagerData}
\alias{sagerQuantData}
\alias{sagerIdData}
\alias{sagerMzMLData}
\alias{sagerGetData}
\title{Sager Data}
\usage{
sagerData(cache, pattern)

sagerQuantData()

sagerIdData()

sagerMzMLData()

sagerGetData(cache = rpx::rpxCache)
}
\arguments{
\item{cache}{Object of class \code{BiocFileCache}. Default for
\code{sagerMzMLData()}, \code{sagerQuantData()} and \code{sagerIdData()} is
\code{\link[rpx:cache]{rpx::rpxCache()}}.}

\item{pattern}{\code{character(1)} defining the pattern used to queray
\code{cache}. These are respectively \code{"sager_suset_PXD016766"},
\code{"sager_subset_quant"} and \code{"sager_subset_id"} for
\code{sagerMzMLData()}, \code{sagerQuantData()} and \code{sagerIdData()}.}
}
\description{
A set of function to install and access test data. Three datasets
are available: a sage TMT11 quantitation tsv file, a sage
identification tsv file, and an mzML that contains the raw data
used as input.

The data are downloaded from ProteomeXchange project PXD016766
(see reference below) and processed with sage version 0.10.0. See
\code{inst/scripts/make-data.R} for details on how these data were
generated and subset.
}
\examples{

sagerQuantData()

sagerIdData()

sagerMzMLData()
}
\references{
Yu \emph{et al.} 'Benchmarking the Orbitrap Tribrid Eclipse for Next
Generation Multiplexed Proteomics' Anal. Chem. 2020, 92, 9,
6478–6485 Publication Date:April 6, 2020
\href{https://doi.org/10.1021/acs.analchem.9b05685}{DOI:10.1021/acs.analchem.9b05685}.
}
