% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sageSummarizedExperiment.R
\name{sageSummarizedExperiment}
\alias{sageSummarizedExperiment}
\title{Import sage results}
\usage{
sageSummarizedExperiment(
  quantFile,
  idFile,
  byQuant = c("file", "scannr"),
  byId = c("filename", "scannr"),
  quantPattern = "tmt_",
  ...
)
}
\arguments{
\item{quantFile}{\code{character(1)} containing the quantification
results, typically "quant.tsv".}

\item{idFile}{\code{character(1)} containing the identification
results, typically "results.sage.tsv".}

\item{byQuant}{\code{character()} containing the specifications of the
quantification columns used for merging. Passed the \code{\link[=merge]{merge()}}
as \code{by.x}. Default is \code{c("file", "scannr")}.}

\item{byId}{\code{character()} containing the specifications of the
identification columns used for merging. Passed the \code{\link[=merge]{merge()}}
as \code{by.y}. Default is \code{c("filename", "scannr")}.}

\item{quantPattern}{\code{character(1)} defining the pattern passed to
\code{\link[=grep]{grep()}} to extract the columns containing quantitative data.}

\item{...}{Additional parameters passed to \code{\link[=read.delim]{read.delim()}}.}
}
\value{
An instance of class \code{\link[=SummarizedExperiment]{SummarizedExperiment()}}.
}
\description{
This function imports sage TMT quantitation and identification
results and combines them into a
\code{\link[SummarizedExperiment:RangedSummarizedExperiment-class]{SummarizedExperiment::SummarizedExperiment()}} object.
}
\examples{

idFile <- BiocFileCache::bfcquery(
                             BiocFileCache::BiocFileCache(),
                             "sageRes")$fpath
quantFile <- BiocFileCache::bfcquery(
                                BiocFileCache::BiocFileCache(),
                                "sageQuant")$fpath

sageSummarizedExperiment(quantFile, idFile)
}
\author{
Laurent Gatto
}
