% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEGR-package.R
\docType{package}
\name{SEGR-package}
\alias{SEGR}
\alias{SEGR-package}
\title{SEGR: What the Package Does (Title Case)}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
