% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Sieges_Voix}
\alias{Sieges_Voix}
\title{Nombre de sièges et de voix dans l'Union européenne}
\format{
Un jeu de données avec 27 observations de 4 variables.
\describe{
\item{Etats.Membres}{}
\item{Date.entrée}{}
\item{Sièges.au.parlement}{}
\item{Voix.au.conseil}{}
}
}
\usage{
Sieges_Voix
}
\description{
Nombre de sièges et de voix dans l'Union européenne.
}
\keyword{datasets}
