% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Total_Pays}
\alias{Total_Pays}
\title{Emploi total par pays dans les pays de l'OCDE en 2020-Q3.}
\format{
Un jeu de données avec 34 observations de 2 variables.
\describe{
\item{NameX}{Acronyme du pays}
\item{Effectif}{Nombres de personnes}
}
}
\usage{
Total_Pays
}
\description{
Emploi total par pays (indicateur). OCDE (2021). \doi{10.1787/6b2fff89-fr}. (Consulté le 11 février 2021).
}
\keyword{datasets}
