% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Total_Secteur}
\alias{Total_Secteur}
\title{Emploi total par secteur dans les pays de l'OCDE en 2020-Q3.}
\format{
Un jeu de données avec 5 observations de 3 variables.
\describe{
\item{Secteur}{Acronyme du secteur d'actvité}
\item{NameX}{Nom du secteur d'actvité}
\item{Effectif}{Nombres de personnes}
}
}
\usage{
Total_Secteur
}
\description{
Emploi total par secteur d'activité (indicateur). OCDE (2021). \href{https://doi.org/10.1787/6b2fff89-fr}{doi: 10.1787/6b2fff89-fr}. (Consulté le 11 février 2021).
L'industrie (\code{INDUSwithoutCONSTR}) désigne l'activité industrielle \emph{SANS} la construction.
}
\keyword{datasets}
