% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Richesse}
\alias{Richesse}
\title{Nombre de personnes dans un foyer}
\format{
Un jeu de données avec 6 observations de 2 variables \code{xi} et \code{ni}.
\describe{
\item{xi}{Nombres de personnes dans un foyer}
\item{ni}{Nombres de foyers}
}
}
\usage{
Richesse
}
\description{
Ménages par taille du ménage en 2017, source INSEE.
}
\keyword{datasets}
