% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagePSM.R
\name{sagePSM}
\alias{sagePSM}
\title{Import sage identification results}
\usage{
sagePSM(
  idTable,
  spectrum = "scannr",
  peptide = "peptide",
  protein = "proteins",
  decoy = "label",
  rank = "rank",
  score = "hyperscore",
  fdr = "spectrum_fdr",
  ...
)
}
\arguments{
\item{idTable}{\code{character(1)} containing the path to the
identification results, typically "results.sage.tsv", or a
\code{data.frame} containing the identification results.}

\item{spectrum}{\code{character(1)} variable name that defines a
spectrum in the PSM data. Default are \code{"scannr"}.}

\item{peptide}{\code{character(1)} variable name that defines a peptide
in the PSM data. Detaults are \code{"peptide"}.}

\item{protein}{protein \code{character(1)} variable name that defines a
protein in the PSM data. Detaults are \code{"proteins"}.}

\item{decoy}{\code{character(1)} variable name that defines a decoy hit
in the PSM data. Detaults are \code{"label"}.}

\item{rank}{\code{character(1)} variable name that defines the
rank of the peptide spectrum match in the PSM data. Default is
\code{"rank"}.}

\item{score}{\code{character(1)} variable name that defines the
PSM score. default is \code{"hyperscore"}.}

\item{fdr}{\code{character(1)} variable name that defines the spectrum
FDR (or any relevant reliability score that can be used for
filtering, such as the PEP) in the PSM data. Default is
\code{"spectrum_fdr"}.}

\item{...}{Additional arguments passed to \code{\link[=read.delim]{read.delim()}}.}
}
\value{
An instance of class \code{\link[=PSM]{PSM()}}.
}
\description{
This function imports sage identification results as a
\code{\link[PSMatch:PSM]{PSMatch::PSM()}} object.
}
\examples{

## Add the data to the package's cache if they
## aren't already available

if (!sagerAvailableData()["id"])
  sagerAddData("id")
basename(f <- sagerIdData())

sagePSM(f)
}
\author{
Laurent Gatto
}
