% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sager-package.R
\docType{package}
\name{sager}
\alias{sager}
\title{Analysing sage results with R}
\description{
The sager package can be used to results produced by the sage
search engine. Sage produces tab-separated output files for
identification and quantitation results. In \code{sager}, we import
these result files into establised Bioconductor classes:
\itemize{
\item Identification results are parsed and imported as
\code{\link[PSMatch:PSM]{PSMatch::PSM()}} objects with \code{\link[=sagerPSM]{sagerPSM()}}.
\item Quantitation (and identification) results are parsed and as
\code{\link[QFeatures:QFeatures-class]{QFeatures::QFeatures()}} objects.
\item It is also possible to import quantitation, identification and
raw spectra into an \code{\link[MsExperiment:MsExperiment]{MsExperiment::MsExperiment()}} object.
}

These functions are described in their respective manual pages
with data generated from Yu \emph{et al.} 'Benchmarking the Orbitrap
Tribrid Eclipse for Next Generation Multiplexed Proteomics',
downloaded from ProteomeXchange project PXD016766. See
\code{\link[=sagerData]{sagerData()}} for details.
}
\references{
\itemize{
\item \strong{Sage}: proteomics searching so fast it seems like
magic. \url{https://github.com/lazear/sage}.
\item Yu \emph{et al.} 'Benchmarking the Orbitrap Tribrid Eclipse for Next
Generation Multiplexed Proteomics' Anal. Chem. 2020, 92, 9,
6478–6485 Publication Date:April 6, 2020
\href{https://doi.org/10.1021/acs.analchem.9b05685}{DOI:10.1021/acs.analchem.9b05685}.
}
}
