% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagerData.R
\name{sager_rids}
\alias{sager_rids}
\alias{sager_urls}
\title{Sager Cache Resource Identifiers}
\usage{
sager_rids(which = c("quant", "id", "mzml"))

sager_urls(which = c("quant", "id", "mzml"))
}
\arguments{
\item{which}{\code{character()} specifying what type of resource to
return.}
}
\description{
\itemize{
\item \code{sager_rids()} is used internally to keep track of the resource
names (rname) in the package's cache. It is used in functions
such as \code{\link[=sagerQuantData]{sagerQuantData()}}, ... that return the path to a
resource (rpath). When new data are added to the cache, their
rname should be added here.
\item \code{sager_urls()} keeps track of the resource URLs on zenodo. It is
used in \code{\link[=sagerAddData]{sagerAddData()}} the get the URLs of the resources to be
added.
}

Both returne named vectors.
}
\examples{

sager_ids()

sager_urls()
}
