% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sagerData}
\alias{sagerData}
\alias{sagerQuantData}
\alias{sagerIdData}
\alias{sagerMzMLData}
\alias{sagerAddData}
\alias{sagerRemoveData}
\title{Sager Data}
\usage{
sagerData(cache, rname)

sagerQuantData()

sagerIdData()

sagerMzMLData()

sagerAddData(
  which = c("quant", "id", "mzml"),
  cache = BiocFileCache::BiocFileCache()
)

sagerRemoveData(
  which = c("quant", "id", "mzml"),
  cache = BiocFileCache::BiocFileCache()
)
}
\arguments{
\item{cache}{Object of class \code{BiocFileCache}. Default for is
\code{\link[BiocFileCache:BiocFileCache-class]{BiocFileCache::BiocFileCache()}}.}

\item{rname}{\code{character(1)} defining the resource name used to
query \code{cache}.}

\item{which}{One or several of \code{"quant"}, \code{"id"} or \code{"mzml"}
specifying what data to download.}
}
\description{
A set of function to install and access test data. Three datasets
are available: a sage TMT11 quantitation tsv file, a sage
identification tsv file, and an mzML file that contains the raw
data used as input.

The data were downloaded from the ProteomeXchange project
PXD016766 (see reference below) and processed with sage version
0.10.0. See \code{inst/scripts/make-data.R} for details on how these
data were generated and subset.
}
\details{
The \code{sagerMzMLData()}, \code{sagerQuantData()} and \code{sagerIdData()} use
the \code{sager::sagerData()} function under the hood, that defined the
resource name and the cache.
}
\section{Functions}{

\itemize{
\item \code{sagerMzMLData()} returns the path to the cached mzML file.
\item \code{sagerQuantData()} returns the path to the cached quantitation
tab-separated file.
\item \code{sagerIdData()} returns the path to the cached identifcation
tab-separated file.
\item \code{sagerAddData(which, cache)} checks if the resources are already available
in the cache (default is \code{\link[BiocFileCache:BiocFileCache-class]{BiocFileCache::BiocFileCache()}}). If not, data are
downloaded and added to the cache.
\item \code{sagerRemoveData(which, cache)} removes the cached resource.
}
}

\examples{

sagerQuantData()

sagerIdData()

sagerMzMLData()
}
\references{
Yu \emph{et al.} 'Benchmarking the Orbitrap Tribrid Eclipse for Next
Generation Multiplexed Proteomics' Anal. Chem. 2020, 92, 9,
6478–6485 Publication Date:April 6, 2020
\href{https://doi.org/10.1021/acs.analchem.9b05685}{DOI:10.1021/acs.analchem.9b05685}.
}
