% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sagerData.R
\name{sagerData}
\alias{sagerData}
\alias{sagerCache}
\alias{sagerQuantData}
\alias{sagerIdData}
\alias{sagerMzMLData}
\alias{sagerAddData}
\alias{sagerRemoveData}
\alias{sagerAvailableData}
\title{Sager Data}
\usage{
sagerData(cache, rname)

sagerCache()

sagerQuantData()

sagerIdData()

sagerMzMLData()

sagerAddData(which = c("quant", "id", "mzml"), cache = sagerCache())

sagerRemoveData(which = c("quant", "id", "mzml"), cache = sagerCache())

sagerAvailableData(which = c("quant", "id", "mzml"), cache = sagerCache())
}
\arguments{
\item{cache}{Object of class \code{BiocFileCache}. Default for is the
package's cache returned by \code{sagerCache()}.}

\item{rname}{\code{character(1)} defining the resource name used to
query \code{cache}.}

\item{which}{One or several of \code{"quant"}, \code{"id"} or \code{"mzml"}
specifying what data to download.}
}
\description{
A set of function to install and access \code{sager} example
data. Three datasets are available: a TMT11 quantitation tsv file,
a identification tsv file, both generated by
\href{https://github.com/lazear/sage}{sage}, and an mzML file that
contains the raw data used as input to sage.

The data were downloaded from the ProteomeXchange project
\href{https://www.ebi.ac.uk/pride/archive/projects/PXD016766}{PXD016766}
and processed with sage version 0.10.0 and subset to reduce their
size. See \code{inst/scripts/make-data.R} for details on how these data
were generated and subset.
}
\section{Functions}{

\itemize{
\item \code{sagerMzMLData()} returns the path to the cached mzML file.
\item \code{sagerQuantData()} returns the path to the cached quantitation
tab-separated file.
\item \code{sagerIdData()} returns the path to the cached identifcation
tab-separated file.
\item \code{sagerAvailableData()} checks if the data are available in the
package cache.
\item \code{sagerAddData(which, cache)} checks if the resources are already
available in the cache. If not, data are downloaded and added
to the cache.
\item \code{sagerRemoveData(which, cache)} removes the cached resource.
\item \code{sagerCache()} returns the package's cache, an object of class
\code{BiocFileCache}.
}

The \code{sagerMzMLData()}, \code{sagerQuantData()} and \code{sagerIdData()} use
the \code{sagerData()} function under the hood, that defined the
resource name and the cache. A user shouldn't have to use
\code{sagerData()}.

The data files are downloaded and cached by \code{sagerAddData()} are
available on zenodo (10.5281/zenodo.7804639).
}

\examples{

## Are data already available
sagerAvailableData()

sagerAvailableData("id")

## Remove data from cache
sagerRemoveData("id")

sagerAvailableData("id")

## Add data (if not already available)
sagerAddData(c("id", "quant"))

## Get the path to the data
sagerQuantData()

sagerIdData()

sagerMzMLData()

## Data are stored in the package's cache
sagerCache()
}
\references{
\itemize{
\item \strong{Sage}: proteomics searching so fast it seems like
magic. \url{https://github.com/lazear/sage}.
\item Project PXD016766:
\href{https://www.ebi.ac.uk/pride/archive/projects/PXD016766}{www.ebi.ac.uk/pride/archive/projects/PXD016766}.
\item Yu \emph{et al.} 'Benchmarking the Orbitrap Tribrid Eclipse for Next
Generation Multiplexed Proteomics' Anal. Chem. 2020, 92, 9,
6478–6485 Publication Date:April 6, 2020
\href{https://doi.org/10.1021/acs.analchem.9b05685}{DOI:10.1021/acs.analchem.9b05685}.
\item sager package test data: \href{https://doi.org/10.5281/zenodo.7804639}{DOI: 10.5281/zenodo.7804639}.
}
}
